Updated LV2 spec.
git-svn-id: http://svn.drobilla.net/lad@602 a436a847-0d15-0410-975c-d299462d15a1

+++ lv2/plugins/osc_print.lv2/lv2.h
@@ -196,7 +196,7 @@ typedef struct _LV2_Descriptor {
    * activate() rather than here.  If a host calls instantiate, it MUST
    * call cleanup() at some point in the future. */
   LV2_Handle (*instantiate)(const struct _LV2_Descriptor *  Descriptor,
-                            uint32_t                        SampleRate,
+                            double                          SampleRate,
                             const char *                    BundlePath,
                             const LV2_Host_Feature *const * HostFeatures);
